Перенесет управление сортировкой из компонента страницы в подчиненый компонен
